Merge redmine_social: Part I
A long time ago eScience was Redmine with heavily customized
redmine_social plugin. Nowadays the core has seen so many modifications
relying on the code of the plugin the split practically isn't there
anymore. Upstream development has ceased to exist pretty much too, so
keeping restoring the split to enable updates of the Redmine core isn't
worth the trouble. So start merging the code of plugin into the core
to enable a more pleasant development expierence. Lots of code was
duplicated in the plugin or introduced as monkey patches.

This commit merges everything under app/ and the patches to to models
and controllers.